+++ code/fwc/src/main/java/com/elementaryengineers/fwc/random/WS_Beg_PicAdd.java
@@ -26,7 +26,7 @@ public class WS_Beg_PicAdd extends WS_Master
     
 //  Constructor  //
     //::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
-    public WS_Beg_PicAdd (long seedValue, int num_fractions, 
+    public WS_Beg_PicAdd (int seedValue, int num_fractions, 
                                            int min_num, int max_num, 
                                            int min_den, int max_den,
                                            int gen_denom_flag,
